requirements.txt
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Session==1.1
googletrans==4.0.0rc1
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mimerender==0.6.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
